backend/requirements.txt
fastapi
uvicorn
pillow
numpy
torch
torchvision
faiss-cpu
pytesseract
opencv-python
pydantic
pydantic-settings
python-multipart
requests